Re-organized the save scalar tests as per frameworks
